Feat/update functions

+++ src/items/items.service.ts
@@ -1,8 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { ItemStatus } from './item-status.enum';
 import { Item } from '../entities/item.entity'; //item.model.tsからitem.entity.tsに変更なんで？
 import { ItemReppository } from './item.repository';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ItemsService {
@@ -21,20 +26,27 @@ export class ItemsService {
     return found;
   }
 
-  async create(CreateItemDto: CreateItemDto): Promise<Item> {
+  async create(CreateItemDto: CreateItemDto, user: User): Promise<Item> {
     //商品オブジェクトの作成はrepositoryに移したのでserviceから削除
-    return await this.ItemRepository.createItem(CreateItemDto);
+    return await this.ItemRepository.createItem(CreateItemDto, user);
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
+    if (item.userId === user.id) {
+      throw new BadRequestException('自身の商品を購入することはできません。');
+    }
     item.updatedAt = new Date().toISOString();
     item.status = ItemStatus.SOLD_OUT;
     await this.ItemRepository.save(item); //オブジェクトの保存はsaveメソッド
     return item;
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません。');
+    }
     await this.ItemRepository.delete({ id });
   }
 }
